+++ src/data/platforms/index.ts
@@ -13,6 +13,24 @@ import { dedao } from './dedao';
 import { xiaoetong } from './xiaoetong';
 import { xiaobaotong } from './xiaobaotong';
 import { afdian } from './afdian';
+import { youtube } from './youtube';
+import { tiktok } from './tiktok';
+import { twitch } from './twitch';
+import { patreon } from './patreon';
+import { buymeacoffee } from './buymeacoffee';
+import { kofi } from './kofi';
+import { substack } from './substack';
+import { ghost } from './ghost';
+import { beehiiv } from './beehiiv';
+import { convertkit } from './convertkit';
+import { gumroad } from './gumroad';
+import { lemonsqueezy } from './lemonsqueezy';
+import { teachable } from './teachable';
+import { kajabi } from './kajabi';
+import { spotifyPodcasters } from './spotify-podcasters';
+import { applePodcasts } from './apple-podcasts';
+import { xTwitter } from './x-twitter';
+import { instagram } from './instagram';
 import type { Platform } from '../types';
 
 export const platforms: Platform[] = [
@@ -31,6 +49,24 @@ export const platforms: Platform[] = [
   xiaoetong,
   xiaobaotong,
   afdian,
+  youtube,
+  tiktok,
+  twitch,
+  patreon,
+  buymeacoffee,
+  kofi,
+  substack,
+  ghost,
+  beehiiv,
+  convertkit,
+  gumroad,
+  lemonsqueezy,
+  teachable,
+  kajabi,
+  spotifyPodcasters,
+  applePodcasts,
+  xTwitter,
+  instagram,
 ];
 
 export {
@@ -49,4 +85,22 @@ export {
   xiaoetong,
   xiaobaotong,
   afdian,
+  youtube,
+  tiktok,
+  twitch,
+  patreon,
+  buymeacoffee,
+  kofi,
+  substack,
+  ghost,
+  beehiiv,
+  convertkit,
+  gumroad,
+  lemonsqueezy,
+  teachable,
+  kajabi,
+  spotifyPodcasters,
+  applePodcasts,
+  xTwitter,
+  instagram,
 };